TASK#14: Name changing 1. Changes name of FilmCard to PromoFilm 2. Changes name of SmallFilmCard to FilmCard
